Categorical Variable Distribution - Output Generation

Using the tabular output generated by cvd_process, this function will build a visualization of the results Each function configuration will output a bespoke ggplot object. Plot theming can be adjusted by the user after the graph has bee output using + theme(). Plots can also be made interactive by using make_interactive_squba() on the object output from cvd_output.

Usage

cvd_output(process_output, filter_concept, vocab_tbl = NULL)

Arguments

process_output

tabular output | The output from cvd_process

filter_concept

numeric/string or vector | The specific code(s) that should be the focus of the analysis; used for ss_anom_la and ms_anom_la

vocab_tbl

tabular input | OPTIONAL: the location of an external vocabulary table containing concept names for the provided codes. if not NULL, concept names will be available in either a reference table or in a hover tooltip

Value

a graph or table to visualize the results from

Examples

#' Source setup file
source(system.file('setup.R', package = 'categoricalvariabledistribution'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'cvd_process_test',
                      working_directory = getwd(),
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = 'extdata',
                      cdm_schema = NA)
#> Connected to: :memory:@NA

#' Build mock study cohort
cohort_tbl <- cdm_tbl('person') %>% dplyr::distinct(person_id) %>%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id %in% c(1:6), 'synth1', 'synth2'))

#' Execute `cvd_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
cvd_process_example <- cvd_process(cohort=cohort_tbl,
                                  domain_tbl=categoricalvariabledistribution::cvd_domain_tbl,
                                  concept_set=categoricalvariabledistribution::cvd_concept_set,
                                  omop_or_pcornet='omop',
                                  multi_or_single_site = 'single',
                                  anomaly_or_exploratory='exploratory',
                                  time = FALSE)
#> Rows: 85 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (5): module, check, Always Required, Required for Check, Optional
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Joining with `by = join_by(person_id, start_date, end_date, site, site_summ, fu_diff, fu)`
#> Error in db_query_fields.DBIConnection(con, ...): Can't query fields.
#> ℹ Using SQL: SELECT * FROM `measurement` AS `q01` WHERE (0 = 1)
#> Caused by error:
#> ! no such table: measurement

cvd_process_example
#> Error: object 'cvd_process_example' not found

#' Execute `cvd_output` function
#' The output was edited for a better indication of what the visualization will
#' look like.
#' The 0s are a limitation of the small sample data set used for this example
cvd_output_example <- cvd_output(process_output = cvd_process_example)
#> Error: object 'cvd_process_example' not found

cvd_output_example
#> Error: object 'cvd_output_example' not found

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(cvd_output_example)
#> Error: object 'cvd_output_example' not found